try to fix format spec for other architectures

+++ zeigen2/src/log.h
@@ -17,7 +17,7 @@ int log_level;
     } while (0);
 
 #define log_section(...) LOG(SECTION, 10, __VA_ARGS__)
-#define log_calloc(amount, size) (9 <= log_level? fprintf(stderr, "CALLOC: %s:%d: \t%s \t%d * %s = %d * %zd\n", __FILE__, __LINE__, __func__, amount, #size, amount, size) : 0)
+#define log_calloc(amount, size) (9 <= log_level? fprintf(stderr, "CALLOC: %s:%d: \t%s \t%d * %s = %d * %zd\n", __FILE__, __LINE__, __func__, amount, #size, amount, (size_t) size) : 0)
 #define log_malloc(size) (8 <= log_level ? fprintf(stderr, "MALLOC: %s:%d: \t%s \t%s = %zd\n", __FILE__, __LINE__, __func__, #size, (size_t) size) : 0)
 #define log_error_in (7 <= log_level ? fprintf(stderr, "ERROR-IN: %s:%d: %s\n", __FILE__, __LINE__, __func__): 0)
 #define log_info(...) LOG(INFO, 6, __VA_ARGS__)
